requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
Pillow>=10.2.0
google-generativeai>=0.7.0